Relocate storages
