+++ src/main/java/com/github/mkram17/bazaarutils/data/stored/BazaarLimitsStorage.java
@@ -1,6 +1,7 @@
-package com.github.mkram17.bazaarutils.utils.storage;
+package com.github.mkram17.bazaarutils.data.stored;
 
 import com.github.mkram17.bazaarutils.features.gui.overlays.BazaarLimitsVisualizer;
+import com.github.mkram17.bazaarutils.utils.storage.ProfileStorage;
 import com.mojang.serialization.Codec;
 
 import java.util.ArrayList;
