+++ src/main/java/com/github/mkram17/bazaarutils/features/gui/overlays/BazaarLimitsVisualizer.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 import com.github.mkram17.bazaarutils.config.features.gui.OverlaysConfig;
 import com.github.mkram17.bazaarutils.utils.codecs.ZonedDateTimeCodec;
-import com.github.mkram17.bazaarutils.utils.storage.BazaarLimitsStorage;
+import com.github.mkram17.bazaarutils.data.stored.BazaarLimitsStorage;
 import com.github.mkram17.bazaarutils.events.BUListener;
 import com.github.mkram17.bazaarutils.generated.BazaarUtilsModules;
 import com.github.mkram17.bazaarutils.misc.BUCompatibilityHelper;
