+++ src/main/java/com/github/mkram17/bazaarutils/utils/bazaar/data/wrappers/CustomBazaarReply.java
@@ -1,11 +1,8 @@
 package com.github.mkram17.bazaarutils.utils.bazaar.data.wrappers;
 
-import com.github.mkram17.bazaarutils.utils.storage.UserOrdersStorage;
-import com.github.mkram17.bazaarutils.utils.bazaar.market.order.Order;
 import lombok.Getter;
 import net.hypixel.api.reply.AbstractReply;
 
-import java.util.List;
 import java.util.Map;
 
 public class CustomBazaarReply extends AbstractReply {
